prompt popup and directory renaming
