+++ foto-social-react/psis-image-based-social-media-app/src/Authentification/Login.tsx
@@ -11,7 +11,7 @@ import {
 } from "@mui/material";
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import ParticleLayer from "../MainPage/ParticleLayer.tsx";
+import ParticleLayer from "../GroupPage/ParticleLayer.tsx";
 
 const styles = {
     lockWrapper: {
