+++ foto-social-react/psis-image-based-social-media-app/src/Authentification/Register.tsx
@@ -11,7 +11,7 @@ import {
 import { LockOutlined } from "@mui/icons-material";
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import ParticleLayer from "../MainPage/ParticleLayer.tsx";
+import ParticleLayer from "../GroupPage/ParticleLayer.tsx";
 
 const styles = {
     lockWrapper: {
