+++ foto-social-react/psis-image-based-social-media-app/src/GroupPage/bottomNavigationBar.tsx
